Move start command to subcommands/start
